Update annotations plugin to use new conventions.
